Refactor restaurant and find page and begin find quiz
