+++ frontend/src/Login.js
@@ -1,5 +1,5 @@
 import './App.css';
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import Form from "react-bootstrap/Form";
 import Cookies from 'universal-cookie';
 import { useHistory } from "react-router-dom";
